broad changes to various components and migrated to SQL server from sqlite

+++ src/Database/Rando.cs
@@ -11,8 +11,8 @@ namespace NinjaBotCore.Database
         public string ListItem { get; set; }
         [Key]
         public Guid ListId { get; set; }
-        public int ListWeight { get; set; }
-        public int NumberOfPlayers { get; set; }
+        public long ListWeight { get; set; }
+        public long NumberOfPlayers { get; set; }
         public Nullable<long> ServerId { get; set; }
         public string AddedBy { get; set; }
         public Nullable<long> SetById { get; set; }
